+++ src/app/page.tsx
@@ -1,8 +1,8 @@
 import Image from 'next/image'
 import styles from './page.module.scss'
-import { MainArticle } from '@/components/main-article/main-article'
-import { Showcase } from '@/components/showcase/showcase'
-import { Guides } from '@/components/guides/guides'
+import { MainArticle } from '@/components/main-article/index'
+import { Showcase } from '@/components/showcase/index'
+import { Guides } from '@/components/guides/index'
 
 export default function Home() {
   return (
