+++ src/components/filter-bar/index.ts
@@ -1,0 +1,1 @@
+export { FilterBar } from './filter-bar'
