+++ src/components/guides/index.ts
@@ -1,0 +1,1 @@
+export { Guides } from './guides'
