+++ src/components/header-bar/index.ts
@@ -1,0 +1,1 @@
+export { HeaderBar } from './header-bar'
