+++ src/components/main-article/index.ts
@@ -1,0 +1,1 @@
+export { MainArticle } from './main-article'
