+++ src/components/showcase/index.ts
@@ -1,0 +1,1 @@
+export { Showcase } from './showcase'
